refac: util 분리 - product, discount
